improvements of the repo.

+++ src/Lib/Parameters/Robot.py
@@ -244,7 +244,7 @@ Universal_Robots_UR3_Str.Theta.Name = [f'Joint_1_{Universal_Robots_UR3_Str.Name}
                                        f'Joint_6_{Universal_Robots_UR3_Str.Name}_ID_{Universal_Robots_UR3_Str.Id:03}']
 Universal_Robots_UR3_Str.Theta.Type = ['R', 'R', 'R', 'R', 'R', 'R']
 Universal_Robots_UR3_Str.Theta.Axis = ['Z', 'Z', 'Z', 'Z', 'Z', 'Z']
-Universal_Robots_UR3_Str.Theta.Direction = np.array([1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.float16)
+Universal_Robots_UR3_Str.Theta.Direction = np.array([1, 1, 1, 1, 1, 1], dtype=np.int8)
 Universal_Robots_UR3_Str.External_Axis = False
 # Colliders of the robot structure.
 #   1\ Internal.
@@ -330,7 +330,7 @@ ABB_IRB_120_Str.Theta.Name = [f'Joint_1_{ABB_IRB_120_Str.Name}_ID_{ABB_IRB_120_S
                               f'Joint_6_{ABB_IRB_120_Str.Name}_ID_{ABB_IRB_120_Str.Id:03}']
 ABB_IRB_120_Str.Theta.Type = ['R', 'R', 'R', 'R', 'R', 'R']
 ABB_IRB_120_Str.Theta.Axis = ['Z', 'Z', 'Z', 'Z', 'Z', 'Z']
-ABB_IRB_120_Str.Theta.Direction = np.array([1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.float16)
+ABB_IRB_120_Str.Theta.Direction = np.array([1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.int8)
 ABB_IRB_120_Str.External_Axis = False
 # Colliders of the robot structure.
 #   1\ Internal.
@@ -421,7 +421,7 @@ ABB_IRB_120_L_Ax_Str.Theta.Name = [f'Joint_L_{ABB_IRB_120_L_Ax_Str.Name}_ID_{ABB
                                    f'Joint_6_{ABB_IRB_120_L_Ax_Str.Name}_ID_{ABB_IRB_120_L_Ax_Str.Id:03}']
 ABB_IRB_120_L_Ax_Str.Theta.Type = ['P', 'R', 'R', 'R', 'R', 'R', 'R']
 ABB_IRB_120_L_Ax_Str.Theta.Axis = ['X', 'Z', 'Z', 'Z', 'Z', 'Z', 'Z']
-ABB_IRB_120_L_Ax_Str.Theta.Direction = np.array([-1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.float16)
+ABB_IRB_120_L_Ax_Str.Theta.Direction = np.array([-1, 1, 1, 1, 1, 1, 1], dtype=np.int8)
 ABB_IRB_120_L_Ax_Str.External_Axis = True
 # Colliders of the robot structure.
 #   1\ Internal.
@@ -524,7 +524,7 @@ ABB_IRB_14000_R_Str.Theta.Name = [f'Joint_1_{ABB_IRB_14000_R_Str.Name}_ID_{ABB_I
                                   f'Joint_6_{ABB_IRB_14000_R_Str.Name}_ID_{ABB_IRB_14000_R_Str.Id:03}']
 ABB_IRB_14000_R_Str.Theta.Type = ['R', 'R', 'R', 'R', 'R', 'R', 'R']
 ABB_IRB_14000_R_Str.Theta.Axis = ['Z', 'Z', 'Z', 'Z', 'Z', 'Z', 'Z']
-ABB_IRB_14000_R_Str.Theta.Direction = np.array([1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.float16)
+ABB_IRB_14000_R_Str.Theta.Direction = np.array([1, 1, 1, 1, 1, 1, 1], dtype=np.int8)
 ABB_IRB_14000_R_Str.External_Axis = False
 # Colliders of the robot structure.
 #   1\ Internal.
@@ -626,7 +626,7 @@ ABB_IRB_14000_L_Str.Theta.Name = [f'Joint_1_{ABB_IRB_14000_L_Str.Name}_ID_{ABB_I
                                   f'Joint_6_{ABB_IRB_14000_L_Str.Name}_ID_{ABB_IRB_14000_L_Str.Id:03}']
 ABB_IRB_14000_L_Str.Theta.Type = ['R', 'R', 'R', 'R', 'R', 'R', 'R']
 ABB_IRB_14000_L_Str.Theta.Axis = ['Z', 'Z', 'Z', 'Z', 'Z', 'Z', 'Z']
-ABB_IRB_14000_L_Str.Theta.Direction = np.array([1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.float16)
+ABB_IRB_14000_L_Str.Theta.Direction = np.array([1, 1, 1, 1, 1, 1, 1], dtype=np.int8)
 ABB_IRB_14000_L_Str.External_Axis = False
 # Colliders of the robot structure.
 #   1\ Internal.
@@ -718,10 +718,10 @@ EPSON_LS3_B401S_Str.Theta.Zero = np.array([0.0, 0.0, 0.0, 0.0],
 EPSON_LS3_B401S_Str.Theta.Home = np.array([Mathematics.Degree_To_Radian(90.0), Mathematics.Degree_To_Radian(0.0), 0.0, Mathematics.Degree_To_Radian(0.0)],
                                           dtype = np.float32)
 # Limits of absolute joint position.
-EPSON_LS3_B401S_Str.Theta.Limit = np.array([[-2.303834612632515, 2.303834612632515], 
+EPSON_LS3_B401S_Str.Theta.Limit = np.array([[  -2.303834612632515,  2.303834612632515], 
                                             [ -2.4609142453120048, 2.4609142453120048], 
-                                            [                0.0,             0.150], 
-                                            [ -3.141592653589793, 3.141592653589793]], dtype = np.float32)
+                                            [                 0.0,              0.150], 
+                                            [  -3.141592653589793,  3.141592653589793]], dtype = np.float32)
 # Parameters of the object (Blender robot arm).
 EPSON_LS3_B401S_Str.Theta.Name = [f'Joint_1_{EPSON_LS3_B401S_Str.Name}_ID_{EPSON_LS3_B401S_Str.Id:03}', 
                                   f'Joint_2_{EPSON_LS3_B401S_Str.Name}_ID_{EPSON_LS3_B401S_Str.Id:03}', 
@@ -729,7 +729,7 @@ EPSON_LS3_B401S_Str.Theta.Name = [f'Joint_1_{EPSON_LS3_B401S_Str.Name}_ID_{EPSON
                                   f'Joint_4_{EPSON_LS3_B401S_Str.Name}_ID_{EPSON_LS3_B401S_Str.Id:03}']
 EPSON_LS3_B401S_Str.Theta.Type = ['R', 'R', 'P', 'R']
 EPSON_LS3_B401S_Str.Theta.Axis = ['Z', 'Z', 'Z', 'Z']
-EPSON_LS3_B401S_Str.Theta.Direction = np.array([1.0, 1.0, -1.0, 1.0], dtype=np.float16)
+EPSON_LS3_B401S_Str.Theta.Direction = np.array([1, 1, -1, 1], dtype=np.int8)
 EPSON_LS3_B401S_Str.External_Axis = False
 # Colliders of the robot structure.
 #   1\ Internal.
